requirements.txt
torchreid
ultralytics==8.0.215
cvzone
tensorboard
gdown
numpy
PyQt5
scipy
scikit-image
dataclasses
opencv-python
Pillow
torch
pydantic
torchvision
typing-extensions
matplotlib
open3d
cvxopt
dvc[gdrive]
pyserial
pyyaml
protobuf
grpcio-tools
pymilvus
python-dotenv
pytorch-lightning